Removing unused method, moving `IndexHelper` to ES module.
