+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/IndexHelper.java
@@ -14,9 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.indexer;
+package org.graylog.storage.elasticsearch6;
 
-import com.google.common.collect.ImmutableSet;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.graylog2.plugin.Message;
@@ -24,29 +23,8 @@ import org.graylog2.plugin.Tools;
 import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
 
 public class IndexHelper {
-    public static Set<String> getOldestIndices(IndexSet indexSet, int count) {
-        final String[] managedIndicesNames = indexSet.getManagedIndices();
-        if (count <= 0 || managedIndicesNames.length <= count) {
-            return Collections.emptySet();
-        }
-
-        final List<Integer> numbers = new ArrayList<>(managedIndicesNames.length);
-        for (String indexName : managedIndicesNames) {
-            indexSet.extractIndexNumber(indexName).ifPresent(numbers::add);
-        }
-
-        final List<String> sorted = prependPrefixes(indexSet.getIndexPrefix(), Tools.asSortedSet(numbers));
-
-        return ImmutableSet.copyOf(sorted.subList(0, count));
-    }
-
     @Nullable
     public static QueryBuilder getTimestampRangeFilter(TimeRange range) {
         if (range == null) {
@@ -56,14 +34,5 @@ public class IndexHelper {
         return QueryBuilders.rangeQuery(Message.FIELD_TIMESTAMP)
                 .gte(Tools.buildElasticSearchTimeFormat(range.getFrom()))
                 .lte(Tools.buildElasticSearchTimeFormat(range.getTo()));
-    }
-
-    private static List<String> prependPrefixes(String prefix, Collection<Integer> numbers) {
-        final List<String> r = new ArrayList<>();
-        for (int number : numbers) {
-            r.add(prefix + "_" + number);
-        }
-
-        return r;
     }
 }
